first, almost empty, version of samsonSetup2
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@1138 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ libs/network/UnhelloedEndpoint.cpp
@@ -191,7 +191,7 @@ Endpoint2::Status UnhelloedEndpoint::helloSend(Message::MessageType type)
 Endpoint2::Status UnhelloedEndpoint::helloExchange(int secs, int usecs)
 {
 	Message::Header       header;
-	int                   dataLen = 0;
+	long                  dataLen = 0;
 	void*                 dataP   = NULL;
 	Message::HelloData*   helloP;
 	Packet                packet(Message::Unknown);
